anichin-fullstack/backend/requirements.txt
requests
bs4
python-dotenv
flask
flask-cors
flask-jwt-extended
flask-bcrypt
flask-pymongo
pymongo
dnspython
gunicorn